chore: separate out cross-browser tests
