+++ packages/kit/test/apps/basics/test/server.test.js
@@ -5,7 +5,7 @@ import { createHash, randomBytes } from 'node:crypto';
 
 /** @typedef {import('@playwright/test').Response} Response */
 
-test.skip(({ javaScriptEnabled }) => javaScriptEnabled);
+test.skip(({ javaScriptEnabled }) => javaScriptEnabled || process.env.KIT_E2E_BROWSER === 'webkit');
 
 test.describe.configure({ mode: 'parallel' });
